% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_F1_db}
\alias{create_F1_db}
\title{Create a "SQLite" F1 database
Create a local 'SQLite' database using the latest 'Ergast' data and establish a connection}
\usage{
create_F1_db(csv_dir = NULL, rm_csv = FALSE)
}
\arguments{
\item{csv_dir}{either NULL or the name of a directory containing csv files from Ergast.
If NULL, the files will be downloaded and placed in a directory within the working directory named "/f1db_csv"}

\item{rm_csv}{logical indicating whether the csv directory should be deleted after initializing the database}
}
\value{
a 'DBI' connection object
}
\description{
Create a "SQLite" F1 database
Create a local 'SQLite' database using the latest 'Ergast' data and establish a connection
}
\details{
create_F1_db() creates a local 'SQLite' database using csv files downloaded from Ergast.
The database will be located in a file "f1_db.sqlite" within the working directory.
}
