% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{createF1db}
\alias{createF1db}
\title{Create an F1 database
Creates a local 'DuckDB' or 'SQLite' database using the latest 'Ergast' data and establish a connection.}
\usage{
createF1db(csv_dir = NULL, rm_csv = FALSE, type = "duckdb")
}
\arguments{
\item{csv_dir}{either NULL or the name of a directory containing csv files from Ergast.
If NULL, the files will be downloaded and placed in a directory within the working directory named "/f1db_csv"}

\item{rm_csv}{logical indicating whether the csv directory should be deleted after initializing the database}

\item{type}{"duckdb" or "sqlite"}
}
\value{
an object of class \code{\link[duckdb:duckdb_connection-class]{duckdb_connection}} or \code{\link[RSQLite:SQLiteConnection-class]{SQLiteConnection}}
}
\description{
Create an F1 database
Creates a local 'DuckDB' or 'SQLite' database using the latest 'Ergast' data and establish a connection.
}
\details{
\code{createF1db()} creates a local 'DuckDB' or 'SQLite' database using csv files downloaded from Ergast.
The database will be located in a file 'f1_db.duckdb' 'f1_db.sqlite' within the working directory.

Databases created with this function can be interacted with using functions from the 'DBI' Package. You can also use the convenience function \code{\link{F1dbConnect}} to reconnect to a database created by \code{createF1db}
}
\examples{
\donttest{
library(DBI)
con <- createF1db()
dbListTables(con)
dbDisconnect(con)
}
}
