% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{downloadErgastCSV}
\alias{downloadErgastCSV}
\title{Download 'Ergast' CSV Database Tables}
\usage{
downloadErgastCSV(destfile = paste0(getwd(), "/f1db_csv"))
}
\arguments{
\item{destfile}{a character string with the name of the directory in which the files are saved}
}
\description{
This function downloads the 'Ergast' database as a set of CSV files and unzips them into a local directory
}
\details{
The CSV files downloaded by this function have column headers and are UTF-8 encoded. Each file contains a database table.
}
