% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{F1dbConnect}
\alias{F1dbConnect}
\title{Connect to an existing F1 SQLite database
Establishes a connection to a SQLite database previously created by \code{\link{createF1db}}.}
\usage{
F1dbConnect(file = "f1_db.sqlite")
}
\arguments{
\item{file}{path to the database file}
}
\value{
an object of class \code{\link[RSQLite:SQLiteConnection-class]{SQLiteConnection}}
}
\description{
Connect to an existing F1 SQLite database
Establishes a connection to a SQLite database previously created by \code{\link{createF1db}}.
}
\examples{
\donttest{
# a file "f1_db.sqlite" already exists in the working directory
con <- F1dbConnect()
circuits <- dbReadTable(con, "circuits")
}
}
