% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{fillMissingQualifyingColumns}
\alias{fillMissingQualifyingColumns}
\title{Detect when a data.frame doesn't have Q2 and Q3 Columns and fill them with NA}
\usage{
fillMissingQualifyingColumns(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Detect when a data.frame doesn't have Q2 and Q3 Columns and fill them with NA
}
