% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getConstructorQualifying}
\alias{getConstructorQualifying}
\title{Get a data.frame of Formula 1 qualifying results for a specific constructor for a given year}
\usage{
getConstructorQualifying(year, constructor)
}
\arguments{
\item{year}{a four digit integer}

\item{constructor}{a string containing an 'Ergast' constructorId, usually the constructor's name in all lowercase}
}
\description{
Get a data.frame of Formula 1 qualifying results for a specific constructor for a given year
}
\examples{
\donttest{mercedes_qualy_2018 <- getConstructorQualifying(2018, "mercedes")}
}
