% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getDriverLaps}
\alias{getDriverLaps}
\title{Get a dataframe of a specific driver's lap times in a Formula 1 Grand Prix}
\usage{
getDriverLaps(year, race, driverId)
}
\arguments{
\item{year}{a four digit integer}

\item{race}{a 1 or 2 digit integer indicating which round of the season}

\item{driverId}{an Ergast driverId, usually the driver's name in all lowercase}
}
\description{
Get a dataframe of a specific driver's lap times in a Formula 1 Grand Prix
}
