% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getF1Schedule}
\alias{getF1Schedule}
\title{Get a data.frame of the Formula 1 schedule for a given year}
\usage{
getF1Schedule(year)
}
\arguments{
\item{year}{a four digit integer or "current"}
}
\description{
Get a data.frame of the Formula 1 schedule for a given year
}
\examples{
\donttest{
schedule_2019 <- getF1Schedule(2020)
}
}
