% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getFinalF1Standings}
\alias{getFinalF1Standings}
\title{Get a data.frame of Formula 1 World Championship final standings for a season}
\usage{
getFinalF1Standings(year, type = "driver")
}
\arguments{
\item{year}{a four digit integer}

\item{type}{constructor or driver}
}
\description{
Get a data.frame of Formula 1 World Championship final standings for a season
}
\examples{
\donttest{driver_standings_2012 <- getFinalF1Standings(2012, type = "driver")
constructor_standings_2012 <- getFinalF1Standings(2012, type = "constructor")}
}
