% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getPitStopsByRace}
\alias{getPitStopsByRace}
\title{Get a dataframe of pit stops in a Formula 1 Grand Prix}
\usage{
getPitStopsByRace(year, race)
}
\arguments{
\item{year}{a four digit integer}

\item{race}{a 1 or 2 digit integer indicating which round of the season}
}
\description{
Get a dataframe of pit stops in a Formula 1 Grand Prix
}
\examples{
\donttest{aus_pitstops_2018 <- getPitStopsByRace(2018, 1)}
}
