% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getResultsByPosition}
\alias{getResultsByPosition}
\title{Get a data.frame of results for a Formula 1 season for a given position}
\usage{
getResultsByPosition(year, position = 1)
}
\arguments{
\item{year}{a four digit integer}

\item{position}{a 1 or 2 digit integer indicating which position to filter results by}
}
\description{
Get a data.frame of results for a Formula 1 season for a given position
}
