% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getSprintResults}
\alias{getSprintResults}
\title{Get a data.frame of results for an F1 Sprint Qualifying session}
\usage{
getSprintResults(year, race)
}
\arguments{
\item{year}{a four digit integer}

\item{race}{a 1 or 2 digit integer indicating which round of the season}
}
\description{
Get a data.frame of results for an F1 Sprint Qualifying session
}
\examples{
\donttest{
# Sprint Qualifying results for the 2021 British Grand Prix
uk_sprint32 <- getSprintResults(2021, 10)
}
}
