% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsf1.R
\name{getQualifyingResults}
\alias{getQualifyingResults}
\title{Get a dataframe of qualifying results for a Formula 1 Grand Prix}
\usage{
getQualifyingResults(year, race)
}
\arguments{
\item{year}{a four digit integer}

\item{race}{a 1 or 2 digit integer indicating which round of the season}
}
\description{
Get a dataframe of qualifying results for a Formula 1 Grand Prix
}
\examples{
aus_qualy_2018 <- getQualifyingResults(2018, 1)
}
